WIP provider refactor
