+++ packages/web3-core-provider/test/unit/detectProviderProtocol.test.ts
