--- packages/web3-core-requestmanager/src/index.ts
@@ -1,61 +1,0 @@
-import Web3ProviderHttp from 'web3-providers-http';
-import {
-    ProviderOptions,
-    RpcResponse,
-    RequestArguments,
-    PollingInfo,
-    IWeb3Provider,
-} from 'web3-providers-base/lib/types';
-
-import { ProviderProtocol } from './types';
-
-export default class Web3RequestManager {
-    provider: IWeb3Provider;
-
-    providerProtocol: ProviderProtocol = ProviderProtocol.UNKNOWN;
-
-    constructor(providerOptions: ProviderOptions) {
-        switch (
-            Web3RequestManager.detectProviderProtocol(
-                providerOptions.providerUrl
-            )
-        ) {
-            case ProviderProtocol.HTTP:
-                this.provider = new Web3ProviderHttp(providerOptions);
-                this.providerProtocol = ProviderProtocol.HTTP;
-                break;
-            case ProviderProtocol.WS:
-                this.providerProtocol = ProviderProtocol.WS;
-                // TODO
-                throw Error('Provider protocol not implemented');
-            case ProviderProtocol.IPC:
-                this.providerProtocol = ProviderProtocol.IPC;
-                // TODO
-                throw Error('Provider protocol not implemented');
-            default:
-                // TODO figure out if possible to support generic provider
-                // could be anything supporting EIP1193's request method
-                throw Error('Provider protocol not supported');
-        }
-    }
-
-    static detectProviderProtocol(providerUrl: string): ProviderProtocol {
-        try {
-            if (/^http(s)?:\/\//i.test(providerUrl)) {
-                return ProviderProtocol.HTTP;
-            } else if (/^ws(s)?:\/\//i.test(providerUrl)) {
-                return ProviderProtocol.WS;
-            } else if (/^ipc:\/\//i.test(providerUrl)) {
-                return ProviderProtocol.WS;
-            }
-            return ProviderProtocol.UNKNOWN;
-        } catch (error) {
-            throw Error(`Error detecting provider protocol: ${error.message}`);
-        }
-    }
-
-    async request(args: RequestArguments): Promise<RpcResponse | PollingInfo> {
-        if (this.provider === undefined) throw Error('No provider initialized');
-        return this.provider.request(args);
-    }
-}
