+++ packages/web3-core-types/.prettierignore
