+++ packages/web3-core-types/README.md
@@ -2,12 +2,12 @@
   <img src="../../assets/logo/web3js.jpg" width="200" alt="web3.js" />
 </p>
 
-# web3.js - Providers Base
+# web3.js - Core Types
 
 ## Installation
 
 ```bash
-yarn add web3-providers-base
+yarn add web3-core-types
 ```
 
 ## Package.json Scripts
