+++ packages/web3-core-types/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "web3-providers-base",
-    "version": "1.0.0-alpha.1",
-    "description": "Ethereum provider base class for Web3 4.x.x",
+    "name": "web3-core-types",
+    "version": "1.0.0-alpha.0",
+    "description": "Various shared types Web3 4.x.x packages",
     "main": "lib/index.js",
     "repository": "https://github.com/ChainSafe/web3.js",
     "author": "ChainSafe Systems",
@@ -29,7 +29,6 @@
         "jest": "^26.6.3",
         "prettier": "^2.3.0",
         "ts-jest": "^26.5.4",
-        "typescript": "^4.2.4",
-        "web3-utils": "4.0.0-alpha.0"
+        "typescript": "^4.2.4"
     }
 }
