+++ packages/web3-eth/package.json
@@ -28,10 +28,12 @@
         "jest": "^26.6.3",
         "prettier": "2.3.0",
         "ts-jest": "^26.5.4",
-        "typescript": "^4.2.4"
+        "typescript": "^4.2.4",
+        "web3-core-types": "1.0.0-alpha.0"
     },
     "dependencies": {
         "web3-core-requestmanager": "4.0.0-alpha.0",
+        "web3-core-provider": "1.0.0-alpha.0",
         "web3-utils": "^4.0.0"
     }
 }
