+++ packages/web3-providers-http/package.json
@@ -29,7 +29,8 @@
         "jest": "^26.6.3",
         "prettier": "^2.3.0",
         "ts-jest": "^26.5.4",
-        "typescript": "^4.2.4"
+        "typescript": "^4.2.4",
+        "web3-core-types": "1.0.0-alpha.0"
     },
     "dependencies": {
         "axios": "^0.21.1",
