+++ packages/web3-providers-http/src/index.ts
@@ -1,28 +1,20 @@
 import axios, { AxiosInstance } from 'axios';
-import Web3ProviderBase from 'web3-providers-base';
 import {
-    ProviderOptions,
     IWeb3Provider,
     RpcResponse,
     RequestArguments,
-    PollingInfo,
-} from 'web3-providers-base/lib/types';
+} from 'web3-core-types/lib/types';
 import { EventEmitter } from 'events';
 
-export default class Web3ProvidersHttp
-    extends Web3ProviderBase
-    implements IWeb3Provider
+export default class Web3ProvidersHttp implements IWeb3Provider
 {
     private _httpClient: AxiosInstance;
-    private _polls: {
-        [pollingId: number]: ReturnType<typeof setTimeout>;
-    } = {};
 
-    constructor(options: ProviderOptions) {
-        super(options);
-        this._httpClient = Web3ProvidersHttp._createHttpClient(
-            options.providerUrl
-        );
+    web3Client: string;
+
+    constructor(web3Client: string) {
+        this._httpClient = Web3ProvidersHttp._createHttpClient(web3Client);
+        this.web3Client = web3Client;
     }
 
     private static _validateProviderUrl(providerUrl: string): boolean {
@@ -46,12 +38,12 @@ export default class Web3ProvidersHttp
         }
     }
 
-    setProvider(providerUrl: string) {
+    setWeb3Client(web3Client: string) {
         try {
-            this._httpClient = Web3ProvidersHttp._createHttpClient(providerUrl);
-            super.providerUrl = providerUrl;
+            this._httpClient = Web3ProvidersHttp._createHttpClient(web3Client);
+            this.web3Client = web3Client;
         } catch (error) {
-            throw Error(`Failed to set provider: ${error.message}`);
+            throw Error(`Failed to set web3 client: ${error.message}`);
         }
     }
 
@@ -59,24 +51,8 @@ export default class Web3ProvidersHttp
         return false;
     }
 
-    async request(args: RequestArguments): Promise<RpcResponse | PollingInfo> {
-        return args.providerOptions?.poll
-            ? this._poll(args)
-            : await this._request(args);
-    }
-
-    cancelPoll(pollingInfo: PollingInfo) {
-        try {
-            if (!this._polls[pollingInfo.pollingId])
-                throw Error(
-                    `Poll with id: ${pollingInfo.pollingId} does not exist`
-                );
-            clearTimeout(this._polls[pollingInfo.pollingId]);
-            pollingInfo.eventEmitter.emit('cancelled');
-            delete this._polls[pollingInfo.pollingId];
-        } catch (error) {
-            throw Error(`Error cancelling poll: ${error.message}`);
-        }
+    async request(args: RequestArguments): Promise<RpcResponse> {
+        return await this._request(args);
     }
 
     private async _request(args: RequestArguments): Promise<RpcResponse> {
@@ -100,37 +76,6 @@ export default class Web3ProvidersHttp
         } catch (error) {
             // TODO Fancy error detection that complies with EIP1193 defined errors
             // https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1193.md#provider-errors
-            throw Error(error.message);
-        }
-    }
-
-    private _poll(args: RequestArguments): PollingInfo {
-        try {
-            if (this._httpClient === undefined)
-                throw Error('No HTTP client initiliazed'); // Is checked in _request, but short circuiting here
-            const pollingInfo = {
-                eventEmitter: new EventEmitter(),
-                pollingId: Math.floor(Math.random() * Number.MAX_SAFE_INTEGER), // generate random integer
-            };
-            this._pollInterval(args, pollingInfo);
-            return pollingInfo;
-        } catch (error) {
-            throw Error(error.message);
-        }
-    }
-
-    private async _pollInterval(
-        args: RequestArguments,
-        pollingInfo: PollingInfo
-    ) {
-        try {
-            const response = await this._request(args);
-            pollingInfo.eventEmitter.emit('response', response);
-            this._polls[pollingInfo.pollingId] = setTimeout(
-                () => this._pollInterval(args, pollingInfo),
-                args.providerOptions?.pollingInterval || 1000
-            );
-        } catch (error) {
             throw Error(error.message);
         }
     }
