+++ packages/web3-utils/package.json
@@ -28,7 +28,8 @@
         "jest": "^26.6.3",
         "prettier": "^2.3.0",
         "ts-jest": "^26.5.4",
-        "typescript": "^4.2.4"
+        "typescript": "^4.2.4",
+        "web3-core-types": "1.0.0-alpha.0"
     },
     "dependencies": {
         "ethereumjs-util": "^7.0.10"
